+++ CHANGELOG.rst
@@ -1,7 +1,7 @@
 Version 0.2.1
 -------------
 
-:Date: TBC
+:Date: 18-September-2022
 
 * Fixing a bug with the serial backend where DTR reset workaround on linux platforms.
   Opening the serial connection was defining encoding as binary rather than the mode.